actualizado

+++ libft/ft_strlen.c
@@ -6,15 +6,15 @@
 /*   By: alegarci <alegarci@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/07 16:10:16 by alegarci          #+#    #+#             */
-/*   Updated: 2025/04/09 17:36:11 by alegarci         ###   ########.fr       */
+/*   Updated: 2025/04/10 15:52:18 by alegarci         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include <unistd.h>
+#include "libft.h"
 
-int	ft_strlen(const char *str)
+size_t	ft_strlen(const char *str)
 {
-	unsigned int	i;
+	size_t	i;
 
 	i = 0;
 	while (str[i])
